rename operator_matrix to operator_state_algebra
