+++ docs/autodoc/unitary_coupled_cluster_util.rst
@@ -7,13 +7,13 @@ Unitary Coupled Cluster Util
 .. automodule:: slowquant.unitary_coupled_cluster.fermionic_operator
     :members:
 
-.. automodule:: slowquant.unitary_coupled_cluster.operator_matrix
+.. automodule:: slowquant.unitary_coupled_cluster.docs/autodoc/unitary_coupled_cluster_util.rst
     :members:
 
 .. automodule:: slowquant.unitary_coupled_cluster.operators
     :members:
 
-.. automodule:: slowquant.unitary_coupled_cluster.operator_extended
+.. automodule:: slowquant.unitary_coupled_cluster.ci_spaces
     :members:
 
 .. automodule:: slowquant.unitary_coupled_cluster.util
