+++ slowquant/unitary_coupled_cluster/ucc_wavefunction.py
@@ -20,7 +20,7 @@ from slowquant.unitary_coupled_cluster.density_matrix import (
     get_electronic_energy,
     get_orbital_gradient,
 )
-from slowquant.unitary_coupled_cluster.operator_matrix import (
+from slowquant.unitary_coupled_cluster.operator_state_algebra import (
     build_operator_matrix,
     construct_ucc_state,
     expectation_value,
